[HAB-177] removed some unused files and moved funcs to differnt files (#83)
