+++ mobile-app/src/atoms/currentUserAtom.ts
@@ -1,6 +1,6 @@
 import { atom } from "jotai";
-import { mockCurrentUser } from "../lib/betterMock";
 import { userWithIdT } from "../lib/db_types";
+import { mockCurrentUser } from "../lib/mockData";
 
 export const currentUserAtom = atom<userWithIdT>(mockCurrentUser);
 export const currentUserIdAtom = atom((get) => get(currentUserAtom).id);
