+++ mobile-app/src/lib/getCurrentUserData.ts
@@ -1,0 +1,6 @@
+import { userWithIdT } from "./db_types";
+import { mockCurrentUser } from "./mockData";
+
+export async function fetchCurentUserInfo(): Promise<userWithIdT> {
+  return mockCurrentUser;
+}
