+++ mobile-app/src/lib/getNumberOfDaysInLastWeek.ts
@@ -1,0 +1,4 @@
+export function getNumberOfDaysInLastWeek() {
+  const currDay = new Date().getDay();
+  return currDay === 0 ? 7 : currDay;
+}
